+++ openbank-kyc-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.10.7](https://github.com/JiRaska/open-bank-oss/compare/kyc-service-v0.10.6...kyc-service-v0.10.7) (2026-08-09)
+
+
+### Bug Fixes
+
+* **libs:** stamp ApiError.timestamp at construction instead of serving 1970 ([#3880](https://github.com/JiRaska/open-bank-oss/issues/3880)) ([b3e6672](https://github.com/JiRaska/open-bank-oss/commit/b3e6672c9e13470fc6353ad8a5483e4075875b1f))
+
 ## [0.10.6](https://github.com/JiRaska/open-bank-oss/compare/kyc-service-v0.10.5...kyc-service-v0.10.6) (2026-08-02)